backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Myounghoon Lee <l.mynghn@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
Pillow = ">=8.3.2"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.2.0"
django-cors-headers = "^3.13.0"
python-dotenv = "^0.20.0"
drf-writable-nested = "^0.7.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.6.0"
ipython = "^8.4.0"
jupyter = "^1.0.0"
httpie = "^3.2.1"
jupyter-contrib-nbextensions = "^0.5.1"
django-extensions = "^3.2.0"
ipdb = "^0.13.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
